pyproject.toml
[project]
name = "oz-main-be-15-team5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aerich>=0.9.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.124.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "tortoise-orm>=0.25.1",
    "uvicorn[standard]>=0.38.0",
    "bcrypt>=4.1.2"
]
# -dev 개발 의존성 추가

[dependency-groups]
dev = [
    "black>=25.12.0",
    "httpx>=0.28.1",
    "mypy>=1.19.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.8",
]
# 배포 의존성 추가

[tool.aerich]
# tortoise ORM 설정이 정의된 파일 경로와 변수 이름
tortoise_config = "app.settings.TORTOISE_CONFIG"
# 마이그레이션 스크립트를 저장할 폴더
location = "./migrations"
